+++ src/asmout.c
@@ -12,11 +12,25 @@
  * Instruction Conversion
  * */
 
+char *convertInstructionAMD_X86_64(AsmOut *out, Instruction ins) {
+	char outBuf[100];
+	if(!strcmp(ins.instruction, "move")) {
+		printf("%s\n", ins.arguments[0]);
+		snprintf(outBuf, sizeof(outBuf), 
+				"\tmov %s(%%rip),%s\n",
+				ins.arguments[0], ins.arguments[1]);
+	}
+	char *ret = (char *)malloc(sizeof(char)*(strlen(outBuf)+2));
+	strcpy(ret, outBuf);
+	return ret;
+} 
+
 char *convertInstructionARM_MAC(AsmOut *out, Instruction ins) {
 	/*
 	 * This is just temporary so I can have something that works...
 	 * */
-	char outBuf[100];
+
+	char outBuf[100] = {0};
 
 	if(!strcmp(ins.instruction, "move")) {
 		/*TODO check instruction argument size for invalid args*/
@@ -32,13 +46,13 @@ char *convertInstructionARM_MAC(AsmOut *out, Instruction ins) {
 				"\tadrp %s,%s@PAGE\n\tadd %s, %s, %s@PAGEOFF\n",
 				"x0", asmVName, "x0", "x0", asmVName);
 	} else if(!strcmp(ins.instruction, "call")) {
-		if(ins.arguments>0) 
-			sprintf(outBuf, "\tbl _%s\n", ins.arguments[0]);
+        if(ins.argLen>0 && ins.arguments[0]!=NULL)
+            snprintf(outBuf, sizeof(outBuf), "\tbl _%s\n", ins.arguments[0]);
 	} else if(!strcmp(ins.instruction, "return")) {
 		sprintf(outBuf, "\tmov x0, #0\n\tbl _exit\n");
 	}
-	char *ret = (char *)malloc(sizeof(char)*(strlen(outBuf)+2));
-	strcpy(ret, outBuf);
+
+	char *ret = strdup(outBuf);
 	return ret;
 }
 
@@ -47,12 +61,12 @@ char *convertInstructionARM_MAC(AsmOut *out, Instruction ins) {
 * */
 
 char *getCPUMain() {
-	char *ret = (char *)malloc(sizeof(char)*15);
+	char *ret = (char *)malloc(sizeof(char)*10);
 	switch(CPU) {
 		case ARM_MAC: strcpy(ret, "main"); break;
-		case AMD_X86_64: strcpy(ret, "start"); break;
+		case AMD_X86_64: strcpy(ret, "main"); break;
+		case I386: strcpy(ret, "main"); break; 
 		case ALPHA: break; /*TODO*/
-		case I386: break; /*TODO*/
 		case ITANIUM_64: break; /*TODO*/
 		case ARMv7: break; /*TODO*/
 		case POWERPC: break; /*TODO*/
@@ -75,49 +89,56 @@ char *createFunctionHeader(char *name) {
 }
 
 
-void convertFunctionsARM_MAC(AsmOut *out) {
+void convertFunctions(AsmOut *out) {
+	out->buffers.functions = (char *)malloc(sizeof(char));
+	/*It actually makes me mad that I'm making sure memory is okay
+	 * because of my schizo ass instruction issues. */
+	*out->buffers.functions = '\0'; 
 	/*x registers are 64-bit w registers are 32*/
 	size_t bufferSize=1;
 	int i,j;
 	for(i=0;i<out->parser->totalFunctions;i++) {
 		char *header = createFunctionHeader(out->parser->functions[i].funName);
-		bufferSize+=(strlen(header)+2);
+		bufferSize+=strlen(header)+1;
 		out->buffers.functions =
-			(char *)realloc(out->buffers.functions, sizeof(char)*bufferSize);
+			realloc(out->buffers.functions, bufferSize);
 		strcat(out->buffers.functions, header);
 		/*instructions*/
 		for(j=0;j<out->parser->functions[i].dataLength;j++) {
-			if(out->parser->functions[i].instructions[j].instruction==NULL) continue;
-			char *asmInstruction = 
-				convertInstructionARM_MAC(out, out->parser->functions[i].instructions[j]);
+			Instruction *curIns = &out->parser->functions[i].instructions[j];
+			if(curIns->instruction==NULL) continue;
+			char *asmInstruction = NULL;
+
+			switch(CPU) {
+				/*alpha*/
+				case ALPHA: break; /*TODO*/
+				/*intel*/
+				case AMD_X86_64:							
+							asmInstruction = convertInstructionAMD_X86_64(out, *curIns);
+							break;
+				case I386: break; /*TODO*/
+				case ITANIUM_64: break; /*TODO*/
+				/*ARM*/
+				case ARM_MAC: 
+							asmInstruction = convertInstructionARM_MAC(out, *curIns);
+							break;
+				case ARMv7: break; /*TODO*/
+				/*IBM*/
+				case POWERPC: break; /*TODO*/
+				case RS6000: break; /*TODO*/
+				case SZ_IBM: break; /*TODO*/
+				/*SUN*/
+				case SPARC: break; /*TODO*/
+			};
+
 			bufferSize+=strlen(asmInstruction)+2;
 			out->buffers.functions =
-				(char *)realloc(out->buffers.functions, sizeof(char)*bufferSize);
+				(char *)realloc(out->buffers.functions, bufferSize);
 			strcat(out->buffers.functions, asmInstruction);
+			free(asmInstruction);
 		}	
 
 	}
-}
-
-void convertFunctions(AsmOut *out) {
-	out->buffers.functions = (char *)malloc(sizeof(char));
-	switch(CPU) {
-		/*alpha*/
-		case ALPHA: break; /*TODO*/
-		/*intel*/
-		case AMD_X86_64: break; /*TODO*/
-		case I386: break; /*TODO*/
-		case ITANIUM_64: break; /*TODO*/
-		/*ARM*/
-		case ARM_MAC: convertFunctionsARM_MAC(out); break;
-		case ARMv7: break; /*TODO*/
-		/*IBM*/
-		case POWERPC: break; /*TODO*/
-		case RS6000: break; /*TODO*/
-		case SZ_IBM: break; /*TODO*/
-		/*SUN*/
-		case SPARC: break; /*TODO*/
-	};	
 }
 
 /* This should be pretty standard along all CPUs
@@ -131,8 +152,7 @@ char *getAsmString(char *name, char *value) {
 			"wl_str.%s:\n\t.asciz %s\n",
 			name, value);
 	strCount++;
-	char *ret = (char *)malloc(sizeof(char)*strlen(buf)+1);
-	strcpy(ret, buf);
+	char *ret = strdup(buf);
 	return ret;
 }
 
@@ -147,8 +167,7 @@ char *getAsmChar(char *name, char *value) {
 			"\n\t.global %s\n%s:\n\t.byte %d\n",
 			nameBuf, nameBuf, value[0]);
 	cCount++;
-	char *ret = (char *)malloc(sizeof(char)*strlen(buf)+1);
-	strcpy(ret, buf);
+	char *ret = strdup(buf);
 	return ret;
 }
 
@@ -162,8 +181,7 @@ char *getAsmInt(char *name, char *value) {
 			"\n\t.global %s\n\t.p2align 2,0x0\n%s:\n\t.long %s\n",
 			nameBuf, nameBuf, hexValue);	
 	iCount++;
-	char *ret = (char *)malloc(sizeof(char)*strlen(buf)+1);
-	strcpy(ret, buf);
+	char *ret = strdup(buf);
 	return ret;
 }